add fixtures to conftest
